+++ app/services/mazeSolver/CheckNodes.scala
@@ -2,7 +2,7 @@ package services.mazeSolver
 import scala.collection.immutable.List
 
 class CheckNodes {
-    def findEntrance(maze: Array[Array[Int]]): (Int, Int) = {
+    def findEntrance(maze: Array[Array[Byte]]): (Int, Int) = {
         for(j <- maze.indices){
             if(maze(j)(0) == 1) {
                 return (j, 0)
@@ -11,7 +11,7 @@ class CheckNodes {
         throw new Exception("Entrance not found; aborting depth first search")
     }
 
-    def checkPossibleNodes(maze: Array[Array[Int]], position: (Int, Int), delta: (Int, Int)): Option[(Int, Int)] ={
+    def checkPossibleNodes(maze: Array[Array[Byte]], position: (Int, Int), delta: (Int, Int)): Option[(Int, Int)] ={
         val (j, i) = position
         val (deltaJ, deltaI) = delta
         try {
@@ -25,7 +25,7 @@ class CheckNodes {
 
     }
 
-    def checkVisitedNodes(visitedNodes: Array[Array[Int]], position: (Int, Int)): Option[(Int, Int)] = {
+    def checkVisitedNodes(visitedNodes: Array[Array[Byte]], position: (Int, Int)): Option[(Int, Int)] = {
         val (j, i) = position
         try {
             if (visitedNodes(j)(i) == 0) {
@@ -37,7 +37,7 @@ class CheckNodes {
         None
     }
 
-    def checkAvailableNodes(maze: Array[Array[Int]], visitedNodes: Array[Array[Int]], position: (Int, Int)): List[(Int, Int)] = {
+    def checkAvailableNodes(maze: Array[Array[Byte]], visitedNodes: Array[Array[Byte]], position: (Int, Int)): List[(Int, Int)] = {
         val deltaList = List((-1, 0), (0, 1), (1, 0), (0, -1))
         val possibleNodes = deltaList.flatMap(delta => checkPossibleNodes(maze, position, delta))
         val availableNodes = possibleNodes.flatMap(possibleNode => checkVisitedNodes(visitedNodes, possibleNode))
